requirements.txt
boto==2.35.1
Collectfast==0.2.1
Django==1.7.5
django-allauth==0.19.0
django-bootstrap-breadcrumbs==0.7.0
django-bootstrap-form==3.1
django-bootstrap3==5.0.3
django-bootstrap3-datetimepicker==2.2.3
django-durationfield==0.5.1
django-extensions==1.5.0
django-jsonfield==0.9.13
django-markdown==0.8.4
django-markdown-deux==1.0.5
django-model-utils==2.2
django-s3-folder-storage==0.2
Django-Select2==4.2.2
django-sizefield==0.7
django-storages==1.1.8
django-tables2==0.15.0
django-uuidfield==0.5.0
easy-thumbnails==2.2
Markdown==2.6
markdown2==2.3.0
oauthlib==0.7.2
phonenumbers==7.0.1
Pillow==2.7.0
psycopg2==2.5.4
python-dateutil==2.4.0
python-openid==2.2.5
pytz==2014.10
requests==2.5.1
requests-oauthlib==0.4.2
six==1.9.0
stripe==1.21.0
